chore: a few more alterations, add a tokenGate for use across routes
