+++ declarations.d.ts
@@ -1,0 +1,1 @@
+type SearchParams = { [key: string]: string | string[] | undefined };
